Move the incremental build mechanism to the launcher with an integrated API
